Use tokio Mutexes that can yield to the executer
The speak() function is now async, and since it needs to borrow the text
to speak over an await, it is easier just to pass the String directly,
we never use it again anyway. You can still pass it an &str that lives
long enough, such as an &'static str.

We also spawn speak() as a separate task, so we can continue to process
events even while speak() waits for a lock.